modified:   client/src/components/Forms/Utils/useValidations.js     added email validation and password validation to file modified:   client/src/components/forms/SignUpOrg.js     removed all old form state and validations to use useForm and useValidation files
